moved notify logic into separate file
